data_analysis_api/requirements.txt
fastapi
uvicorn
python-dotenv
crewai
langchain-google-genai
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
mysql-connector-python
